Moved the core abstractions into /core to help enforce encapsulation boundaries moving forward.
