+++ example/lib/spikes/editor_abstractions/core/composition/document_composer_actions.dart
@@ -1,11 +1,16 @@
-import 'package:example/spikes/editor_abstractions/document/document_editor.dart';
-import 'package:example/spikes/editor_abstractions/document/document_nodes.dart';
-import 'package:example/spikes/editor_abstractions/document/rich_text_document.dart';
-import 'package:example/spikes/editor_abstractions/selectable_text/selectable_text.dart';
-import 'package:example/spikes/editor_abstractions/layout/document_layout.dart';
-import 'package:example/spikes/editor_abstractions/selection/editor_selection.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+
+import '../document/rich_text_document.dart';
+import '../document/document_editor.dart';
+import '../layout/document_layout.dart';
+import '../selection/editor_selection.dart';
+
+// TODO: should nodes be in core?
+import '../document/document_nodes.dart';
+
+// TODO: get rid of this import
+import '../../selectable_text/selectable_text.dart';
 
 class ComposerKeyboardAction {
   const ComposerKeyboardAction.simple({
