+++ example/lib/spikes/editor_abstractions/core/document/document_editor.dart
