+++ example/lib/spikes/editor_abstractions/core/document/rich_text_document.dart
