+++ example/lib/spikes/editor_abstractions/core/selection/editor_selection.dart
@@ -1,8 +1,9 @@
-import 'package:example/spikes/editor_abstractions/document/document_nodes.dart';
-import 'package:example/spikes/editor_abstractions/document/rich_text_document.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+
+import '../document/document_nodes.dart';
+import '../document/rich_text_document.dart';
 
 /// A selection within a `RichTextDocument`.
 ///
